WIP: Add favorite icon to vessel name header and refactor component
