+++ frontend/src/domain/shared_slices/Vessel.js
@@ -343,6 +343,7 @@ const vesselSlice = createSlice({
         afterDateTime: null,
         beforeDateTime: null
       }
+      state.isFocusedOnVesselSearch = false
       state.tripMessagesLastToFormerDEPDateTimes = []
 
       if (!atLeastOneVesselSelected(state.vesselsTracksShowed, state.selectedVesselIdentity)) {
